moved detect_auto_generated_files

+++ benchmark/qsort/c_build.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 
-rm -rf bin
-mkdir -p bin
-cd bin
+option=${1:-"build"}
 
-cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=1 \
-      -DCMAKE_C_COMPILER=clang \
-      -DCMAKE_CXX_COMPILER=clang++ \
-      -DCMAKE_C_FLAGS="-finstrument-functions -g -gdwarf-4" \
-      -DCMAKE_CXX_FLAGS="-finstrument-functions -g -gdwarf-4" ..
+if [[ "$option" == "init" ]]; then
+      rm -rf bin
+      mkdir -p bin
+      cd bin
+
+      cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=1 \
+            -DCMAKE_C_COMPILER=clang \
+            -DCMAKE_CXX_COMPILER=clang++ \
+            -DCMAKE_C_FLAGS="-finstrument-functions -g -gdwarf-4" \
+            -DCMAKE_CXX_FLAGS="-finstrument-functions -g -gdwarf-4" ..
+
+else
+    cd bin
+fi
+
 cmake --build .
